# swagger.yml
swagger: "2.0"
info:
  version: "0.0.1"
  title: "Account Manager  System"
host: "localhost:9090"
basePath: "/v1"
schemes:
  - "http"
consumes:
  -  "application/json"
produces:
  - "application/json"
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /user/{user_id}:
    get:
      operationId: "getUser"
      description: "get user base on id"
      tags:
        - service
      parameters:
        - in: path
          type: string
          required: true
          name: user_id
      responses:
        200:
          description: "successfully save user object into database"
          schema:
            $ref: '#/definitions/user_response'
        401:
          $ref: '#/responses/401'
        404:
          $ref: '#/responses/404'
        default:
          $ref: '#/responses/500'
    delete:
      operationId: "deleteUser"
      description: "delete user"
      tags:
        - service
      parameters:
        - in: path
          name: user_id
          type: string
          required: true
      responses:
        204:
          description: "successfully delete user"
        401:
          $ref: '#/responses/401'
        404:
          $ref: '#/responses/404'
        default:
          $ref: '#/responses/500'
    patch:
      operationId: "updateUser"
      description: "update user name or address"
      tags:
        - service
      parameters:
        - in: body
          name: user
          required: true
          schema:
            $ref: '#/definitions/update_user'
        - in: path
          name: user_id
          type: string
          required: true
      responses:
        200:
          description: "successfully update user"
          schema:
              $ref: '#/definitions/user_response'
        401:
          $ref: '#/responses/401'
        404:
          $ref: '#/responses/404'
        default:
          $ref: '#/responses/500'
definitions:
  user_response:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      email:
        type: string
      address:
        type: string
      created_at:
        type: string
  update_user:
    type: object
    properties:
      name:
        type: string
        minLength: 3
        maxLength: 32
      address:
        type: string
...
  error:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: string
        example: error code
      message:
        type: string
        example: error message
responses:
...
  401:
    description: Unauthorized
    schema:
      $ref: '#/definitions/error'
...
  404:
    description: Not Found
    schema:
      $ref: '#/definitions/error'
...
  500:
    description: Internal Server Error
    schema:
      $ref: '#/definitions/error'